set up test for basic MIDI connectivity
